# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AprovaFácil API",
    "description": "API para sistema de estudos e preparação para concursos",
    "version": "1.1.0",
    "contact": {
      "name": "AprovaFacil Team",
      "email": "contato@aprovafacil.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.aprovafacil.com",
      "description": "Servidor de produção"
    }
  ],
  "paths": {
...
    "/v1/simulados/{slug}/questoes": {
      "get": {
        "operationId": "listarQuestoesPorSlugV1",
        "tags": [
          "Simulados"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questões do simulado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "ETag da seção de questões",
                "schema": {
                  "type": "string"
                }
              },
              "Last-Modified": {
                "description": "Timestamp da última modificação das questões",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "X-Request-Id": {
                "description": "ID de correlação/requisição",
                "schema": {
                  "type": "string"
                }
              },
              "X-Server-Duration": {
                "description": "Duração do processamento no servidor (ms)",
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
...
    "/simulados/slug/{slug}/questoes": {
      "get": {
        "operationId": "getQuestoesBySimuladoSlug",
        "summary": "Listar questões do simulado por slug",
        "description": "Retorna as questões de um simulado identificado por slug.",
        "tags": [
          "Simulados"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug do simulado",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de questões",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "simulado_id": {
                            "type": "string"
                          },
                          "numero_questao": {
                            "type": "integer"
                          },
                          "enunciado": {
                            "type": "string"
                          },
                          "alternativas": {
                            "type": "object"
                          },
                          "resposta_correta": {
                            "type": "string"
                          },
                          "explicacao": {
                            "type": "string"
                          },
                          "disciplina": {
                            "type": "string"
                          },
                          "assunto": {
                            "type": "string"
                          },
                          "dificuldade": {
                            "type": "string"
                          },
                          "ordem": {
                            "type": "integer"
                          },
                          "peso_disciplina": {
                            "type": "integer"
                          },
                          "ativo": {
                            "type": "boolean"
                          },
                          "criado_em": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "atualizado_em": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Simulado não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/questoes-semanais": {
      "get": {
        "operationId": "getQuestoesSemanais",
        "summary": "Listar questões semanais",
        "description": "Retorna lista de questões semanais disponíveis",
        "tags": [
          "Questões Semanais"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "concurso",
            "in": "query",
            "description": "Filtrar por concurso",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "disciplina",
            "in": "query",
            "description": "Filtrar por disciplina",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "semana",
            "in": "query",
            "description": "Número da semana",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 52
            }
          },
          {
            "name": "ano",
            "in": "query",
            "description": "Ano de referência",
            "schema": {
              "type": "integer",
              "minimum": 2020
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de questões semanais",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "questoes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "questao-semanal-123"
                          },
                          "titulo": {
                            "type": "string",
                            "example": "Questão Semanal #15"
                          },
                          "description": {
                            "type": "string",
                            "example": "Questão de Direito Constitucional"
                          },
                          "concurso": {
                            "type": "string",
                            "example": "OAB"
                          },
                          "disciplina": {
                            "type": "string",
                            "example": "Direito Constitucional"
                          },
                          "semana": {
                            "type": "integer",
                            "example": 15
                          },
                          "ano": {
                            "type": "integer",
                            "example": 2024
                          },
                          "dataPublicacao": {
                            "type": "string",
                            "format": "date"
                          },
                          "status": {
                            "type": "string",
                            "example": "disponivel"
                          },
                          "tentativas": {
                            "type": "integer",
                            "example": 0
                          },
                          "acertos": {
                            "type": "integer",
                            "example": 0
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postQuestoesSemanais",
        "summary": "Criar questão semanal",
        "description": "Cria uma nova questão semanal",
        "tags": [
          "Questões Semanais"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "titulo",
                  "concurso",
                  "disciplina",
                  "enunciado",
                  "alternativas",
                  "respostaCorreta"
                ],
                "properties": {
                  "titulo": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200,
                    "description": "Título da questão"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Descrição da questão"
                  },
                  "concurso": {
                    "type": "string",
                    "description": "Concurso relacionado"
                  },
                  "disciplina": {
                    "type": "string",
                    "description": "Disciplina da questão"
                  },
                  "semana": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 52,
                    "description": "Número da semana"
                  },
                  "ano": {
                    "type": "integer",
                    "minimum": 2020,
                    "description": "Ano de referência"
                  },
                  "enunciado": {
                    "type": "string",
                    "minLength": 10,
                    "description": "Enunciado da questão"
                  },
                  "alternativas": {
                    "type": "array",
                    "minItems": 2,
                    "maxItems": 5,
                    "items": {
                      "type": "string"
                    },
                    "description": "Alternativas da questão"
                  },
                  "respostaCorreta": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Índice da resposta correta"
                  },
                  "explicacao": {
                    "type": "string",
                    "description": "Explicação da resposta"
                  },
                  "nivel": {
                    "type": "string",
                    "enum": [
                      "facil",
                      "medio",
                      "dificil"
                    ],
                    "default": "medio",
                    "description": "Nível de dificuldade"
                  }
                }
              },
              "examples": {
                "questao": {
                  "summary": "Exemplo de questão semanal",
                  "value": {
                    "titulo": "Questão Semanal #15 - Direito Constitucional",
                    "description": "Questão sobre princípios fundamentais",
                    "concurso": "OAB",
                    "disciplina": "Direito Constitucional",
                    "semana": 15,
                    "ano": 2024,
                    "enunciado": "Qual é o princípio fundamental da República Federativa do Brasil?",
                    "alternativas": [
                      "A soberania",
                      "A cidadania",
                      "A dignidade da pessoa humana",
                      "O pluralismo político"
                    ],
                    "respostaCorreta": 2,
                    "explicacao": "A dignidade da pessoa humana é o fundamento da República",
                    "nivel": "medio"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Questão semanal criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Questão semanal criada com sucesso"
                    },
                    "questao": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "questao-semanal-456"
                        },
                        "titulo": {
                          "type": "string",
                          "example": "Questão Semanal #15"
                        },
                        "dataCriacao": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Mensagem de erro"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              },
              "message": {
                "type": "string",
                "example": "Mensagem de erro"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}